requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
